Add setup.py for Setuptools support
Very preliminary effort here, though I don't presently know what more
would actually be done.  Do note that this breaks old .yaml files
since they include references like
`!!python/object:library.canmatrix.Frame`.

This can 'easily' be recursively fixed for all YAML files via:

    find . -name '*.yaml' -exec sed -i \
    's/object:library.canmatrix/object:canmatrix.canmatrix/g' {} \;

 #17
