+++ TODO.md
@@ -4,12 +4,12 @@
 
 
 
-library/canmatrix.py
+canmatrix/canmatrix.py
 : Definitions should be imported with disassembling not as complete string
 
 
 
-library/canmatrixGenerateJS.py
+canmatrix/canmatrixGenerateJS.py
  wrong bits are used if signal is not
 
 
@@ -26,7 +26,7 @@ library/canmatrixGenerateJS.py
 
 
 
-library/exportarxml.py
+canmatrix/exportarxml.py
  Well, ..., this is the first attempt to export a arxml-file; I did this without reading any spec;
  recievers of signals are missing
 : insert Speed - if possible
@@ -41,6 +41,8 @@ library/exportarxml.py
 
 
 
+canmatrix/exportkcd.py
+ : support for extended frames (https://github.com/julietkilo/CANBabel/blob/master/src/main/java/com/github/canbabel/canio/kcd/Message.java) :
 
 
 
@@ -62,25 +64,25 @@ library/exportarxml.py
 
 
 
-library/importarxml.py
+canmatrix/importarxml.py
  Well, ..., this is the first attempt to import a arxml-file; I did this without reading any spec;
     : use diagAddress for frame-classification
                     
 
 
 
-library/importdbc.py
+canmatrix/importdbc.py
  support for: VERSION, NS, BS_, SIG_VALTYPE_, BA_DEF_REL == BA_DEF_??, BA_DEF_DEF_REL_ = BA_DEF_DEF_  ??
 
 
 
-library/importdbf.py
+canmatrix/importdbf.py
  support for [START_PARAM_NODE_RX_SIG]
  support for [START_PARAM_NODE_TX_MSG]
 
 
 
-library/importkcd.py
+canmatrix/importkcd.py
  baudrate missing
  name save
  defaults for CAN-Simulation missing
@@ -88,12 +90,15 @@ library/importkcd.py
 
 
 
-library/importsym.py
+canmatrix/importsym.py
  : the signalname should be unique. If same signal-name is used in different mux-ids, the signal-name should be unified. maybe this is an dbc-export-issue not an sym-import-issue
  : I don't know what to do with the "/p"-Options in the signal-definition
  : what is the differenze between SEND and SENDRECEIVE ?
 : support for title example: Title="some title"
                              /p ???
+
+
+
