+++ canmatrix/compare.py
