+++ canmatrix/copy.py
