+++ canmatrix/importxlsx.py
@@ -21,23 +21,96 @@
 #DAMAGE.
 
 #
-# this script imports excel-files to a canmatrix-object
+# this script imports excel-files (xlsx) to a canmatrix-object
 # these Excelfiles should have following collums:
 # ID, Frame Name, Cycle Time [ms], Launch Type, Launch Parameter, Signal Byte No., Signal Bit No., Signal Name, Signal Function,  Signal Length [Bit], Signal Default, Signal Not Available, [LIST OF ECUS], Value,     Name / Phys. Range,     Function / Increment Unit
 #
 
 from __future__ import division
-from __future__ import print_function
 from builtins import *
 import math
-from library.canmatrix import *
-import xlrd
+from .canmatrix import *
 import codecs
+import zipfile
+from xml.etree.ElementTree import iterparse
 
-def importXls(filename):
-    wb = xlrd.open_workbook(filename, formatting_info=True)
-    sh = wb.sheet_by_index(0)
+def readXlsx( fileName, **args ):
+    #from: Hooshmand zandi http://stackoverflow.com/a/16544219
+    import zipfile
+    from xml.etree.ElementTree import iterparse
+
+    if "sheet" in args:
+        sheet=args["sheet"]
+    else:
+        sheet=1
+    if "header" in args:
+        isHeader=args["header"]
+    else:
+        isHeader=False
+
+    rows   = []
+    row    = {}
+    header = {}
+    z      = zipfile.ZipFile( fileName )
+
+    # Get shared strings
+    strings = [ el.text for e, el
+                        in  iterparse( z.open( 'xl/sharedStrings.xml' ) )
+                        if el.tag.endswith( '}t' )
+                        ]
+    value = ''
+
+    # Open specified worksheet
+    for e, el in iterparse( z.open( 'xl/worksheets/sheet%d.xml'%( sheet ) ) ):
+        # get value or index to shared strings
+        if el.tag.endswith( '}v' ):                                   # <v>84</v>
+            value = el.text
+        if el.tag.endswith( '}c' ):                                   # <c r="A3" t="s"><v>84</v></c>
+            # If value is a shared string, use value as an index
+
+            if el.attrib.get( 't' ) == 's':
+                value = strings[int( value )]
+
+            # split the row/col information so that the row leter(s) can be separate
+            letter = el.attrib['r']                                   # AZ22
+            while letter[-1].isdigit():
+                letter = letter[:-1]
+
+            # if it is the first row, then create a header hash for the names
+            # that COULD be used
+            if rows ==[]:
+                header[letter]=value.strip()
+            else:
+                if value != '':
+
+                    # if there is a header row, use the first row's names as the row hash index
+                    if isHeader == True and letter in header:
+                        row[header[letter]] = value
+                    else:
+                        row[letter] = value
+
+            value = ''
+        if el.tag.endswith('}row'):
+            rows.append(row)
+            row = {}
+    z.close()
+    return [header, rows]
+
+def getIfPossible(row, value):
+    if value in row:
+        return row[value].strip()
+    else:
+        return None
+
+def importXlsx(filename):
+    sheet = readXlsx( filename, sheet = 1, header = True )
     db = CanMatrix()
+    letterIndex = []
+    for a in 'ABCDEFGHIJKLMNOPQRSTUVWXYZ':
+        letterIndex.append(a)
+    for a in 'ABCDEFGHIJKLMNOPQRSTUVWXYZ':
+        for b in 'ABCDEFGHIJKLMNOPQRSTUVWXYZ':
+            letterIndex.append("%s%s" % (a,b))
 
     #Defines not imported...
 #       db.addBUDefines("NWM-Stationsadresse",  'HEX 0 63')
@@ -52,71 +125,78 @@ def importXls(filename):
     db.addSignalDefines("GenSigSNA", 'STRING')
 
     # eval search for correct collums:
-    index = {}
-    for i in range(sh.ncols):
-        value = sh.cell(0,i).value
-        if  value == "ID":
-            index['ID'] = i
-        elif "Frame Name" in value:
-            index['frameName'] = i
-        elif "Cycle" in value:
-            index['cycle'] = i
-        elif "Launch Type" in value:
-            index['launchType'] = i
-        elif "Launch Parameter" in value:
-            index['launchParam'] = i
-        elif "Signal Byte No." in value:
-            index['startbyte'] = i
-        elif "Signal Bit No." in value:
-            index['startbit'] = i
-        elif "Signal Name" in value:
-            index['signalName'] = i
-        elif "Signal Function" in value:
-            index['signalComment'] = i
-        elif "Signal Length" in value:
-            index['signalLength'] = i
-        elif "Signal Default" in value:
-            index['signalDefault'] = i
-        elif "Signal Not Ava" in value:
-            index['signalSNA'] = i
-        elif "Value" in value:
-            index['Value'] = i
-        elif "Name / Phys" in value:
-            index['ValueName'] = i
-        elif "Function /" in value:
-            index['function'] = i
-        elif "Byteorder" in value:
-            index['byteorder'] = i
-
-    if "byteorder" in index:
-        index['BUstart'] = index['byteorder'] + 1
+#       index = {}
+#       for i in range(sh.ncols):
+#               value = sh.cell(0,i).value
+#               if  value == "ID":
+#                       index['ID'] = i
+#               elif "Frame Name" in value:
+#                       index['frameName'] = i
+#               elif "Cycle" in value:
+#                       index['cycle'] = i
+#               elif "Launch Type" in value:
+#                       index['launchType'] = i
+#               elif "Launch Parameter" in value:
+#                       index['launchParam'] = i
+#               elif "Signal Byte No." in value:
+#                       index['startbyte'] = i
+#               elif "Signal Bit No." in value:
+#                       index['startbit'] = i
+#               elif "Signal Name" in value:
+#                       index['signalName'] = i
+#               elif "Signal Function" in value:
+#                       index['signalComment'] = i
+#               elif "Signal Length" in value:
+#                       index['signalLength'] = i
+#               elif "Signal Default" in value:
+#                       index['signalDefault'] = i
+#               elif "Signal Not Ava" in value:
+#                       index['signalSNA'] = i
+#               elif "Value" in value:
+#                       index['Value'] = i
+#               elif "Name / Phys" in value:
+#                       index['ValueName'] = i
+#               elif "Function /" in value:
+#                       index['function'] = i
+#               elif "Byteorder" in value:
+#                       index['byteorder'] = i
+    if 'Byteorder' in list(sheet[0].values()):
+        for key in sheet[0]:
+            if sheet[0][key].strip() == 'Byteorder':
+                _BUstart = letterIndex.index(key)+1
+                break
     else:
-        index['BUstart'] = index['signalSNA'] + 1
-    index['BUend'] = index['Value']
+        for key in sheet[0]:
+            if sheet[0][key].strip() == 'Signal Not Available':
+                _BUstart = letterIndex.index(key)+1
+
+    for key in sheet[0]:
+        if sheet[0][key].strip() == 'Value':
+            _BUend = letterIndex.index(key)
 
     #BoardUnits:
-    for x in range(index['BUstart'],index['BUend']):
-        db._BUs.add(BoardUnit(sh.cell(0,x).value))
+    for x in range(_BUstart,_BUend):
+        db._BUs.add(BoardUnit(sheet[0][letterIndex[x]]))
 
     #initialize:
     frameId = None
     signalName = ""
     newBo = None
 
-    for rownum in range(1,sh.nrows):
+    for row in sheet[1]:
         #ignore empty row
-        if sh.cell(rownum,index['ID']).value.__len__() == 0:
-            break
+        if not 'ID' in row:
+            continue
         # new frame detected
-        if sh.cell(rownum,index['ID']).value != frameId:
+        if row['ID'] != frameId:
             sender = []
             # new Frame
-            frameId = sh.cell(rownum,index['ID']).value
-            frameName = sh.cell(rownum,index['frameName']).value
-            cycleTime = sh.cell(rownum,index['cycle']).value
-            launchType = sh.cell(rownum,index['launchType']).value
+            frameId = row['ID']
+            frameName = row['Frame Name']
+            cycleTime = getIfPossible(row,"Cycle Time [ms]")
+            launchType = getIfPossible(row,'Launch Type')
             dlc = 8
-            launchParam = sh.cell(rownum,index['launchParam']).value
+            launchParam = getIfPossible(row,'Launch Parameter')
             if type(launchParam).__name__ != "float":
                 launchParam = 0.0
             launchParam = str(int(launchParam))
@@ -124,7 +204,7 @@ def importXls(filename):
             newBo = Frame(int(frameId[:-1], 16), frameName, dlc, None)
             db._fl.addFrame(newBo)
 
-            #eval launctype
+            #eval launchtype
             if launchType is not None:
                 if "Cyclic+Change" == launchType:
                     newBo.addAttribute("GenMsgSendType", "5")
@@ -147,33 +227,32 @@ def importXls(filename):
                     newBo.addAttribute("GenMsgSendType", "1")
                     newBo.addAttribute("GenMsgDelayTime", launchParam)
 
-            #eval cycletime
+#                       #eval cycletime
             if type(cycleTime).__name__ != "float":
                 cycleTime = 0.0
             newBo.addAttribute("GenMsgCycleTime", str(int(cycleTime)))
 
         #new signal detected
-        if sh.cell(rownum,index['signalName']).value != signalName:
+        if row['Signal Name'] != signalName:
             # new Signal
             reciever = []
-            startbyte = int(sh.cell(rownum,index['startbyte']).value)
-            startbit = int(sh.cell(rownum,index['startbit']).value)
-            signalName = sh.cell(rownum,index['signalName']).value
-            signalComment = sh.cell(rownum,index['signalComment']).value.strip()
-            signalLength = int(sh.cell(rownum,index['signalLength']).value)
-            signalDefault = sh.cell(rownum,index['signalDefault']).value
-            signalSNA = sh.cell(rownum,index['signalSNA']).value
+            startbyte = int(row["Signal Byte No."])
+            startbit = int(row['Signal Bit No.'])
+            signalName = row['Signal Name']
+            signalComment = getIfPossible(row,'Signal Function')
+            signalLength = int(row['Signal Length [Bit]'])
+            signalDefault = getIfPossible(row,'Signal Default')
+            signalSNA = getIfPossible(row,'Signal Not Available')
             multiplex = None
-            if signalComment.startswith('Mode Signal:'):
+            if signalComment is not None and signalComment.startswith('Mode Signal:'):
                 multiplex = 'Multiplexor'
                 signalComment = signalComment[12:]
-            elif signalComment.startswith('Mode '):
+            elif signalComment is not None and signalComment.startswith('Mode '):
                 mux, signalComment = signalComment[4:].split(':',1)
                 multiplex = int(mux.strip())
 
-            if "byteorder" in index:
-                signalByteorder = sh.cell(rownum,index['byteorder']).value
-
+            signalByteorder = getIfPossible(row,'Byteorder')
+            if signalByteorder is not None:
                 if 'i' in signalByteorder:
                     byteorder = 1
                 else:
@@ -184,11 +263,14 @@ def importXls(filename):
             valuetype = '+'
 
             if signalName != "-":
-                for x in range(index['BUstart'],index['BUend']):
-                    if 's' in sh.cell(rownum,x).value:
-                        newBo.addTransmitter(sh.cell(0,x).value.strip())
-                    if 'r' in sh.cell(rownum,x).value:
-                        reciever.append(sh.cell(0,x).value.strip())
+                for x in range(_BUstart,_BUend):
+                    buName = sheet[0][letterIndex[x]].strip()
+                    buSenderReceiver = getIfPossible(row,buName)
+                    if buSenderReceiver is not None:
+                        if 's' in buSenderReceiver:
+                            newBo.addTransmitter(buName)
+                        if 'r' in buSenderReceiver:
+                            reciever.append(buName)
                 if signalLength > 8:
                     newSig = Signal(signalName, (startbyte-1)*8+startbit, signalLength, byteorder, valuetype, 1, 0, 0, 1, "", reciever, multiplex)
                 else:
@@ -196,11 +278,11 @@ def importXls(filename):
 
                 newBo.addSignal(newSig)
                 newSig.addComment(signalComment)
-                function = sh.cell(rownum,index['function']).value
-        value = str(sh.cell(rownum,index['Value']).value)
-        valueName = sh.cell(rownum,index['ValueName']).value
+                function = getIfPossible(row, 'Function / Increment Unit')
+        value = getIfPossible(row,'Value')
+        valueName = getIfPossible(row,'Name / Phys. Range')
 
-        if valueName == 0:
+        if valueName == 0 or valueName == None:
             valueName = "0"
         elif valueName == 1:
             valueName = "1"
@@ -210,18 +292,15 @@ def importXls(filename):
         factor = 0
         unit = ""
 
-        factor = sh.cell(rownum,index['function']).value
-        if type(factor).__name__ == "unicode" or type(factor).__name__ == "str" :
+        factor = getIfPossible(row,'Function / Increment Unit')
+        if type(factor).__name__ == "unicode" or  type(factor).__name__ == "str":
             factor = factor.strip()
             if " " in factor and factor[0].isdigit():
                 (factor, unit) = factor.strip().split(" ",1)
                 factor = factor.strip()
                 unit = unit.strip()
                 newSig._unit = unit
-                try:
-                    newSig._factor = float(factor)
-                except:
-                    print("Some error occured while decoding scale: Signal: %s; \"%s\"" % (signalName, sh.cell(rownum,index['function']).value))
+                newSig._factor = float(factor)
             else:
                 unit = factor.strip()
                 newSig._unit = unit
@@ -242,7 +321,7 @@ def importXls(filename):
 
 
         elif valueName.__len__() > 0:
-            if value.strip().__len__() > 0:
+            if value is not None and value.strip().__len__() > 0:
                 value = int(float(value))
                 newSig.addValues(value, valueName)
             maxi = pow(2,signalLength)-1
